+++ yash-cli/tests/scripted_test/export-y.sh
@@ -5,6 +5,11 @@ export foo-bar=1
 echo not reached
 __IN__
 
+test_O -d -e n 'export rejects variable name starting with a digit' -o portable
+export 1abc=1
+echo not reached
+__IN__
+
 test_OE -e 0 'export accepts non-portable variable name without the portable option'
 export foo-bar=1
 __IN__